analyzer: move http helper clients out of evmnfts
